requirements.txt
Django==5.0
django-utils-six==2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1